requirements.txt
streamlit
python-dotenv
faiss-gpu-cu12
langchain-community
langchain-core
langchain
nltk
groq
llm-guard
langchain-groq
pandas
PyMuPDF
chardet
pdfplumber
spacy
pypdf
sentence-transformers
